Booking API exceptions ADDED

+++ pages/api/booking/[id].js
@@ -14,7 +14,9 @@ const handler = async (req, res) => {
       const booking = await Booking.findById(id);
       res.status(200).json(booking);
     } catch (err) {
-      res.status(500).json(err);
+      res
+        .status(500)
+        .json(err, { message: "The desired booking wasn't found." });
     }
   }
   if (method == "PUT") {
@@ -24,15 +26,23 @@ const handler = async (req, res) => {
       });
       res.status(200).json(booking);
     } catch (err) {
-      res.status(500).json(err);
+      res
+        .status(500)
+        .json({ message: "An error occurred while updating the booking." });
     }
   }
   if (method === "DELETE") {
     try {
+      const booking = await Booking.findById(id);
+      if (!booking) {
+        return res.status(404).json({ message: "Booking not found." });
+      }
       await Booking.findByIdAndDelete(id);
       res.status(200).json({ message: "Successfully deleted the booking" });
     } catch (err) {
-      res.status(500).json(err);
+      res
+        .status(500)
+        .json(err, { message: "The desired booking wasn't found." });
     }
   }
 };
